build: split `Config` into a separate module
